service/storage_proxy: move cordinator_query_options to coordinator_query.hh
It will be soon used by code which doesn't want to include
`service/storage_proxy.hh` just for this.
The move is mostly mechanical, some `service::` has to be sprinkled
around.
