+++ cdc/log.cc
@@ -1673,7 +1673,7 @@ public:
         const auto select_cl = adjust_cl(write_cl);
 
       try {
-        return _ctx._proxy.query(_schema, std::move(command), std::move(partition_ranges), select_cl, service::storage_proxy::coordinator_query_options(default_timeout(), empty_service_permit(), client_state)).then(
+        return _ctx._proxy.query(_schema, std::move(command), std::move(partition_ranges), select_cl, coordinator_query_options(default_timeout(), empty_service_permit(), client_state)).then(
                 [s = _schema, partition_slice = std::move(partition_slice), selection = std::move(selection)] (coordinator_query_result qr) -> lw_shared_ptr<cql3::untyped_result_set> {
             return make_lw_shared<cql3::untyped_result_set>(*s, std::move(qr.query_result), *selection, partition_slice);
         });
